# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  utils.SuccessResponse:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@shiftley.com
    name: API Support
    url: http://www.shiftley.com/support
  description: Shiftley Backend API Documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Shiftley API
  version: "1.0"
paths:
...
  /employees/me/pay-fine:
    post:
      description: Clears the user's debt (Mock Payment).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Pay Unpaid Fines
      tags:
      - Employee
  /employers/me:
    get:
      description: Fetches employer details and business metrics.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Employer Dashboard Profile
      tags:
      - Employer
  /employers/me/gigs:
    get:
      description: Fetches employer's gig history with status filtering.
      parameters:
      - description: Gig Status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: View My Gigs (History & Active)
      tags:
      - Employer
  /employers/me/subscription:
    post:
      consumes:
      - application/json
      description: Generates a mock Razorpay Order ID for subscription purchase.
      parameters:
      - description: Plan Details
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Purchase Subscription Plan
      tags:
      - Employer
  /employers/profiles/employees/{empId}:
    get:
      description: Fetches detailed resume and past feedback of a worker.
      parameters:
      - description: Employee User ID
        in: path
        name: empId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: View Full Worker Profile
      tags:
      - Employer
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"